Add files via upload
In my last submission, I did not create new variables for my final graph. This is a minor oversight and does not ultimately affect the graphs as sequentially the previous graph is already printed after the variables are reused. I changed the variables to be more consistent in this version. 

Thanks.